move entry point for dispatcher into ui folder, update usages
