--- app/src/lib/dispatcher/index.ts
@@ -1,2 +1,0 @@
-export * from './dispatcher'
-export * from './error-handlers'
