+++ app/src/ui/add-repository/add-existing-repository.tsx
@@ -1,7 +1,7 @@
 import { remote } from 'electron'
 import * as React from 'react'
 
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { isGitRepository } from '../../lib/git'
 import { isBareRepository } from '../../lib/git'
 import { Button } from '../lib/button'
