+++ app/src/ui/app-menu/app-menu-bar-button.tsx
@@ -3,7 +3,7 @@ import { IMenu, ISubmenuItem } from '../../models/app-menu'
 import { MenuListItem } from './menu-list-item'
 import { AppMenu, CloseSource } from './app-menu'
 import { ToolbarDropdown } from '../toolbar'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 
 interface IAppMenuBarButtonProps {
   /**
