+++ app/src/ui/app-menu/app-menu.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { MenuPane } from './menu-pane'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { IMenu, MenuItem, ISubmenuItem } from '../../models/app-menu'
 import { SelectionSource, ClickSource } from '../lib/list'
