+++ app/src/ui/app.tsx
@@ -11,7 +11,7 @@ import {
   SuccessfulMergeBannerState,
   MergeConflictsBannerState,
 } from '../lib/app-state'
-import { Dispatcher } from '../lib/dispatcher'
+import { Dispatcher } from './dispatcher'
 import { AppStore, GitHubUserStore, IssuesStore } from '../lib/stores'
 import { assertNever } from '../lib/fatal-error'
 import { shell } from '../lib/app-shell'
