+++ app/src/ui/changes/oversized-files-warning.tsx
@@ -5,7 +5,7 @@ import { Dialog, DialogContent, DialogFooter } from '../dialog'
 import { LinkButton } from '../lib/link-button'
 import { Monospaced } from '../lib/monospaced'
 import { PathText } from '../lib/path-text'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { Repository } from '../../models/repository'
 import { ICommitContext } from '../../models/commit'
 import { DefaultCommitMessage } from '../../models/commit-message'
