+++ app/src/ui/delete-branch/delete-pull-request-dialog.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 
 import { Repository } from '../../models/repository'
 import { Branch } from '../../models/branch'
