+++ app/src/ui/dispatcher/index.ts
@@ -1,0 +1,2 @@
+export * from '../../lib/dispatcher/dispatcher'
+export * from '../../lib/dispatcher/error-handlers'
