+++ app/src/ui/history/merge-call-to-action-with-conflicts.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { MergeResultStatus, HistoryTabMode } from '../../lib/app-state'
 import { Repository } from '../../models/repository'
 import { Branch } from '../../models/branch'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { Button } from '../lib/button'
 import { MergeStatusHeader } from './merge-status-header'
 import { MergeResultKind } from '../../models/merge'
