+++ app/src/ui/merge-conflicts/merge-conflicts-dialog.tsx
@@ -3,7 +3,7 @@ import { join } from 'path'
 import { Button } from '../lib/button'
 import { ButtonGroup } from '../lib/button-group'
 import { Dialog, DialogContent, DialogFooter } from '../dialog'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { PopupType } from '../../models/popup'
 import { RepositorySectionTab } from '../../lib/app-state'
 import { Repository } from '../../models/repository'
