+++ app/src/ui/missing-repository.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import { UiView } from './ui-view'
-import { Dispatcher } from '../lib/dispatcher'
+import { Dispatcher } from './dispatcher'
 import { Repository } from '../models/repository'
 
 import { Button } from './lib/button'
