+++ app/src/ui/publish-repository/publish.tsx
@@ -3,7 +3,7 @@ import {
   PublishRepository,
   IPublishRepositorySettings,
 } from './publish-repository'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { Account } from '../../models/account'
 import { Repository } from '../../models/repository'
 import { ButtonGroup } from '../lib/button-group'
