+++ app/src/ui/repository-settings/repository-settings.tsx
@@ -4,7 +4,7 @@ import { Remote } from './remote'
 import { GitIgnore } from './git-ignore'
 import { assertNever } from '../../lib/fatal-error'
 import { IRemote } from '../../models/remote'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { PopupType } from '../../models/popup'
 import { Repository } from '../../models/repository'
 import { Button } from '../lib/button'
