+++ app/src/ui/sign-in/sign-in.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import {
   SignInState,
   SignInStep,
