+++ app/src/ui/welcome/sign-in-dot-com.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { WelcomeStep } from './welcome'
 import { SignIn } from '../lib/sign-in'
-import { Dispatcher } from '../../lib/dispatcher'
+import { Dispatcher } from '../dispatcher'
 import { SignInState } from '../../lib/stores'
 import { Button } from '../lib/button'
